pyproject.toml
[tool.poetry]
name = "self-learning project"
version = "0.1.0"
description = ""
authors = ["Stas <stas_sever@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "5.0.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
django-cors-headers = "^4.4.0"
drf-yasg = "^1.21.7"
pillow = "^10.4.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-django = "^4.8.0"
pytest-cov = "^5.0.0"
flake8 = "^7.1.1"
flake8-pyproject = "^1.2.3"
factory-boy = "^3.3.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests/*/test_*.py"]
addopts = "--cov=. --cov-report=term-missing --tb=short --cov-config=pyproject.toml"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
omit = [
    "tests/*",
    "config/*",
    "manage.py",
    "*/__init__.py",
    "*/migrations/*"
]

[tool.flake8]
max-line-length = 120
exclude = ["*/migrations/*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
